Changed Chart loading so the user can select charts or directories containing charts or both.
pom.xml
Changed to use Java 1.7

InstallManager.java
Altered the message printed at the to of the Chart tab to advise the user that he can select charts or directories or both.
Changed the chartFileChooser to accept FILES_AND_DIRECTORIES
Changed the ChartFileChooser approveSelection method to handle the multiple selections and construct a List of files that will be processed.